Even more proper ratelimit implementation
We can now task dispatch events without the fear of user abuse.

Flow is as follows:
1. When request is made a call is made to a GenServer.
2. GenServer checks if call would be ratelimited
  * If no, execute request, update buckets based on headers, return to
  caller
  * If yes, start new task, storing the caller
3. New task waits for the timeout period to expire.
4. Call is made again to the GenServer -> GOTO 1, the from variable is
updated throughout so we can return to the original caller

The GenServer is living in its own process and can only process one
request at a time. This ensures that every request that comes in checks
against previous headers, executes, and then updates the headers before
the next request comes in.

To check for ratelimiting, I'm storing ratelimit headers in an ETS table.
When a request comes in, I check to see if there are remaining requests
for the specified endpoint and if there's currently a global lock. These
values dictate how long the task waits before trying to call again.