Break everything
I don't think it's worth the effort to try and force everything into a
struct.

When dealing with json it's not a problem, but the websocket
information comes in as a term and the only trivial way of turning that
into nested structs is to convert it to json and back, negating the
benefits of using etf in the first place.

We're still going to keep the structs to define the atoms we expect to see
from the API, but everything will be a vanilla map from here on out, until
I decide to change my mind again