Added `Snowflake.from_datetime/1` (#90)
* Added `Snowflake.from_datetime/1`

This lets the user generate snowflakes on their own. This is useful when the user wants to generate an artificial snowflake for grabbing messages at a certain point in time.

* Added `discord_epoch` attribute to `Snowflake` module

Before this commit, we were using a magic number in these functions. Magic numbers are bad, and should be given a name.

* Removed piping operator

* Changed from_datetime to return ok error tuples

This gives the user more flexibility for error checking. We also add a banged version for user convenience.

* Moved discord epoch constant to Constants module

In case library users want to use the snowflake, and to keep the epoch constant in a more suitable module.

* Added `no_return` to creation_time spec

Because it could happen, we should document it.

* Removed `no_return` from `creation_time/1`

Since we're no longer using a banged function in the implementation, we can remove the `no_return`.

* Added in min max tests for `creation_time/1`

We do this to test if the function can break at are minimum and maximum ranges on the snowflake. Since the tests pass, our current spec on the `creation_time/1` is safe.