Improve Gateway Error Handling
get_new_gateway_url/0 was not matching the cases of Api.request/3
correctly, which resulted in CaseClauseError being thrown when real
errors were occurring beneath.

Added a less restrictive case that bubbles up an ok/error tuple to
gateway/0 and allows it to make a decision on how to handle the error.

gateway/0 now raises the returned error message if get_new_gateway_url/0 fails.