Do not cast all incoming keys to atoms.
This is only adjusted in the session and voice session dispatcher. That said,
it should be compatible with the current code: atoms are transformed in most
`to_struct` methods via `maybe_to_atom` anyways. In order to fix an atom leak,
the following changes are introduced additionally:

- Adding new keys from the API to interaction structs.
- Adding the `ApplicationCommandInteractionDataResolved` struct.

This may or may not break some things. I will test it with
[bolt](https://github.com/jchristgit/bolt) to look out for any problems, but if
I inferred everything correctly, it would work just fine with existing code.

Note that the new way of converting incoming data to structs also allows us to
slowly phase out `maybe_to_atom`, with the drawback that we need to specify
parsed keys manually.

Fixes #252.