Fix key error on guild update.
Discord sends us string keys here, including the `"id"` key which we use
for looking up the guild in the cache. To properly upsert this, we use
the proper key and cast the new guild to a guild struct before merging
it with the old guild.

This also fixes us not returning the old guild on update, which the
consumer wants.

Fixes #290.