Revert to normalizing gateway keys to atoms.
By the past bug reports and various issues related to variation of key
types coming from the gateway, this has become more of a hassle than
not.

This commit goes back to the behaviour of normalizing all keys to atoms,
with the small exception that snowflakes which look like integers are
parsed as integers. This was based off Matt Nowack (@ihumanable)s work
in our Discord channel, with a small documentation addition. Thank you!

Struct fields getters are updated to `.key` or `[:key]` depending on
whether or not they are documented as optional in the Discord API
documentation, with the exception of the `PartialApplication.flags`
field, which I blandly assume to be present.

Fixes #290.
Supersedes #291.
Supersedes #299.

Co-authored-by: Matt Nowack <matt@discordapp.com>