Implement pluggable presence caching.
Based off `pluggable-guild-cache` to prevent merge conflicts.

This uses the same approach as the pluggable guild cache, in that we
have a base behaviour and dispatcher for the presence cache, and a
ETS-based implementation chosen by default which uses the current
implementation. A small test is added to ensure correct cache function.