Simplify bangify methods in guild cache.
Guild caches provide banged methods such as `get!/1`, `select!/2` and
friends, which were implemented directly in the backing guild cache. It
makes more sense to define these once in the dispatcher to prevent
having to duplicate logic across cache backends. This is already being
performed in the pluggable user cache.