Add information to Api Rest only section
While using Nostrum with Rest Only approach, an error occurred in runtime of build (released with mix release):
```
[info] Application example exited: exited in: Example.Application.start(:normal, [])
    ** (EXIT) an exception was raised:
        ** (ArgumentError) The module Nostrum.Api.Ratelimiter was given as a child to a supervisor but it does not exist.
            (elixir 1.12.2) lib/supervisor.ex:631: Supervisor.init_child/1
```

The cause of it related to all `runtime: false` deps are not included into build, while building with `mix release`.
I decided to add this info to docs as a time saver for devs, who will meet this non-obvious problem only in production.