Document and typespec allowed mentions in webhooks
Closes #440.

Tested locally, with documentation copied from the previous version.
However, testing revealed that passing in a list currently breaks:

    iex(4)> Nostrum.Api.execute_webhook(webhook_id, webhook_token, %{content: "hello @everyone", allowed_mentions: [{:user, user_id}]})
    ** (BadMapError) expected a map, got: {:user, user_id}
        (elixir 1.14.2) lib/map.ex:344: Map.put_new/3
        (nostrum 0.6.1) lib/nostrum/api.ex:4132: Nostrum.Api.prepare_allowed_mentions/1
        (nostrum 0.6.1) lib/nostrum/api.ex:3035: Nostrum.Api.execute_webhook/4
        iex:4: (file)

This may need some further investigation if this functionality is
needed. Or, alternatively, we remove our preparation of
`:allowed_mentions` entirely and rely on the user to pass in
`:allowed_mentions` in Discord's expected format.