Replace GenStage with `:pg`
Previously, the behaviour of GenStage in terms of buffering events and
causing memory bloat as seen in #427 were a bit unclear to determine.
Testing `:min_demand` and `:max_demand` did not seem to bring much
success here, as we still copied around our events between plenty of
processes.

Instead of the multiple-step GenStage, this change will manage the cache
in the shard session to prevent accumulation of more items while we're
still ingesting something else, and then use `:pg` to dispatch events to
consumers. This also now allows to implement multiple consumers that
each receive the events instead of the current round-robin fashion with
GenStage.

This change bumps our OTP requirement from 22 to 23+. This is on Debian
stable at time of writing, so I think we can consider it stable enough.