Prevent crash with retry_after
First off, a retry_after time of exactly `0` is also a "you can go now":
even if it represents some minimal value like 0.001, the network latency
will eat that up anyways. Next up, if the ratelimiter asks us to retry
and we do, handle the case where another process managed to race us and
we need to sit it out again, with up to three retries.

Closes #493.