Convert ratelimiter into an async state machine
This commit reimplements the existing ratelimiter GenServer using a
state machine via `:gen_statem`. This brings us the following benefits:

- Requests are pipelined as soon as we know that the ratelimit allows us
  to. For any endpoint that returns a body, this happens as soon as we
  have retrieved the request headers, even before we have started
  parsing the body.

- No more garbage collection of ratelimiter buckets is necessary. They
  are automatically removed as soon as any pending retry-after headers
  have expired.

- Clients no longer need to sleep out ratelimits. The server does not
  sleep anything out either. In fact, the server is a fully-asynchronous
  hellbeast that just pipelines in the next request as soon as it sees
  fit.

- To be more specific, queueing requests does not require anybody to
  look up anything aside from the bucket, because the bucket's queue
  will effectively handle all queueing as soon as possible.

- The special handling of the global ratelimit has mostly been removed,
  individual requests don't need to look that up at all. The global
  ratelimit is implemented as a special state in the state machine that
  postpones anything you ask it to do until the global ratelimit has
  expired again.

- The impure ratelimit bucket store using four different callbacks to
  user-defined data stores has been removed and replaced with what's
  effectively `:erlang.send_timer`, 30 years of sane language design,
  and a small bit of magic powder.

- The previous "ask gun to reconnect forever and don't do anything on
  down or up events" has been removed, so we can properly tell clients
  when their request is done.

- We can use `:sys.trace(Nostrum.Api.Ratelimiter, true)` or other
  functions from `:sys` to get really valuable insights into what the
  ratelimiter is doing.

This is not just any ratelimiter state machine, this is _The_ State
Machine. This is The Way going forward.

As for the drawbacks: well, distributing the ratelimiter is a bit more
difficult now. And if something very bad happens to the ratelimiter
while a request is in flight, that request is gonna have a real bad
time. Previously we had the request sleeps on client-side so that was
mostly fine. I also have not implemented any back-off logging (yet?).

As for alternatives - I have not really found an alternative which
doesn't involve responding to a specific process that can deal with
ratelimit responses _or_ using some external data storage for it. And
depending on how you do it, that brings you synchronisation problems.

> of course you throw away the first few implementations, you didn't
> understand the problem when you wrote those!