Join ConsumerGroup synchronously
At the moment there is a race going on between when the consumer starts
and when the shard dispatches the first events. For bots that do things
like loading commands on READY, sometimes they may not do it at all as
the consumer is not subscribed and running yet when the shard comes up.
A proper fix for this would be starting the Consumer before the shard,
but with the current architecture of having the consumer in the client
app but not the shard, no synchronized startup procedure is possible
here (unless we add some form of
make-the-shard-wait-for-the-consumer-to-tell-it-that-it-is-up feature,
which I would claim is a bit hacky).