A proof of concept for auto generated bangs.
This can also be applied to async functions. eg take the function
edit_message. adding @unsafe and @async attributes to the function
can handle generating edit_message! and edit_message_async respectively.

This only contains the code for the banged versions however should
the implementaion be approved the async code is trivial.

This also allows all the manual bang functions to be remove from the api
module. (Approx 70)