Remove blob of zlib data from state machine errors
Due to how zlib and the state machine work, when the state machine
crashed with an error, it would log the last blob that was received.

This blob could not be decoded from error traces because it depends on
the context that it was received by, additionally we cannot decode it as
an error logging step as once it has been decoded by a context it cannot
be inflated again.

Removing this blob is the best solution here as it shrinks the size of
state machine errors and makes them easier to read by users and Nostrum
team.

In state machine traces, the payload will now be shown as "PAYLOAD
REMOVED", if the last event was not a payload, the event will remain
unmodified.