Check for youtube-dl version on application startup
Recent updates to YouTube break old versions of youtube-dl, including
the last released version from 2021.

We add a startup check to attempt to fetch the version using the
youtube-dl command line if the configured youtube-dl path contains
`youtube-dl` (to avoid running this check for `yt-dlp`).

The warning advises the user to switch to a maintained fork and outputs
the located youtube-dl binary and the path it is at.

To give the user flexibility, this warning can be disabled using another
configuration key which will suppress the check from running.