Add support for zstd-stream compression
This adds support for zstd-stream compression through configuration of
the :gateway_compression configuration option.

This will optionally use the :ezstd library to decompress received
payloads from Discord allowing for some potential savings on payload
size.

This also moves the ws_state to be agnostic and references the
compression context as `compress_ctx` instead of the previous
`zlib_ctx`.