Optimize mnesia cache querying
QLC queries via mnesia-based caches that would use the `{traverse,
{select, MatchSpec}}` in any shape or form would cause the QLC query to
be executed in two parts, the `mnesia:table` call running the entire
table over the selected match specification, and then another Erlang
list comprehension that would filter the results of that across a list
comprehension in plain Erlang. Per the discussion in #622, this is how
such a query would look like under `qlc:info/1`:

    qlc:q([
           Guild ||
               {GuildId, Guild} <-
                   mnesia:table(nostrum_guilds,
                                [{n_objects, 100},
                                 {lock, read} |
                                 {traverse,
                                  {select,
                                   [{{'_', '$1', '$2'},
                                     [],
                                     [{{'$1', '$2'}}]}]}}]),
               GuildId =:= RequestedGuildId
          ])

The issue is that neither QLC nor mnesia can cleanly optimize it here:
Mnesia does not know about the condition specified in QLC, and QLC's
optimization to use a `lookup_fun` is knocked out by the fact that it
can't reach into the traverse call to detect the reordered columns. It
might be possible to implement this in QLC itself, given it is smart
enough to figure out when to use the lookup function based on the
indices, together with some cooperation from Mnesia itself. Obviously,
this behaviour would lead to unacceptable performance.

This commit introduces an optimization that allows guild, member and
presence cache implementations to export a `query_handle/1` function
that accepts a match specification guard, that is, the "middle part" of
a match specification. The guard determines which rows shall be
filtered. Do note, however, that this is still unable to perform a
complete optimization of lookups of single records - it will still
traverse the table, but in the native ETS code.