Move to being started under the user's supervisor
Usage via `Nostrum.Application` is removed, but the application still
starts the vital checks. To facilitate this, the new module
`Nostrum.Bot` is added.

The goal here is to change the current way users start nostrum.
Presently, it is still only possible to launch a single bot due to our
reliance on globals.

The `%WSState{}` struct is expanded to include the bot token and the
consumer module that should be used. This is kind of an implementation
detail. It would be better if we would keep `%WSState{}` the internal
state struct, and add a new sub-struct that tracks user-relevant
information which is also passed to the consumers.

Reviewing the code, it is also apparent that 1. Erlang is very efficient
at spawning processes, and 2. we should maybe consider optimizing
consumer event dispatch, because at startup we get such a flurry of
events that it's all a bit suboptimal.